+++ lib/glimmer/data_binding/model_binding.rb
@@ -150,7 +150,7 @@ module Glimmer
           end
           @computed_observer_collection.delete(observer)
         elsif nested_property?
-          remove_nested_observers(observer)
+          # No need to call remove_nested_observers(observer) (cleanup happens automatically indirectly when invoked through observer.unobserve(model_binding))
           nested_property_observers_for(observer).clear
         else
           observer.unobserve(model, property_name)
@@ -203,25 +203,6 @@ module Glimmer
         end
       end
       
-      def remove_nested_observers(observer)
-        nested_property_observers = nested_property_observers_for(observer)
-        Concurrent::Array.new(nested_models.zip(nested_property_names)).each_with_index do |zip, i|
-          model, property_name = zip
-          nested_property_observer = nested_property_observers[property_name]
-          previous_index = i - 1
-          if previous_index.negative?
-            parent_model = self
-            parent_property_name = nil
-            parent_observer = observer
-          else
-            parent_model = nested_models[previous_index]
-            parent_property_name = nested_property_names[previous_index]
-            parent_observer = nested_property_observers[parent_property_name]
-          end
-          parent_observer&.unregister_all_observables
-        end
-      end
-
       def call(value, *extra_args)
         return if model.nil?
         converted_value = value